backend/requirements.txt
fastapi==0.100.0
websockets==11.0.3
sqlalchemy==2.0.19
python-dotenv==1.0.0
alembic==1.14.0
uvicorn[standard]
openai[standard]